refactor(core): Move `typeorm` operators from `WaitTracker` to `ExecutionRepository` (no-changelog)
